# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API 接口文档, 一个适用于 Java 程序员的 Golang web 模板",
        "title": "Go_Web_Template",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:10086",
    "basePath": "/api",
    "paths": {
        "/user/admin/create": {
            "post": {
                "description": "管理员接口, 使用默认密码 12345678 创建一个用户, 返回用户 id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "admin"
                ],
                "summary": "创建新用户",
                "parameters": [
                    {
                        "description": "创建用户参数",
                        "name": "userCreateDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-uint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/admin/delete/{id}": {
            "delete": {
                "description": "管理员接口, 删除用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "admin"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要删除的用户 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/admin/info/{id}": {
            "get": {
                "description": "管理员接口, 根据 id 获取用户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "admin"
                ],
                "summary": "查询用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询的用户 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-vo_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/admin/page/{curPage}/{pageSize}": {
            "post": {
                "description": "管理员接口, 分页获取用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "admin"
                ],
                "summary": "查询用户列表",
                "parameters": [
                    {
                        "description": "用户查询条件",
                        "name": "user",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "curPage",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-entity_Page-vo_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/admin/update": {
            "put": {
                "description": "管理员接口, 更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "admin"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "用户更新参数",
                        "name": "userUpdateDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "从 session 中获取已登录用户的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取登录用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-vo_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "接收用户账号密码进行登录, 并保存登录态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录参数",
                        "name": "userLoginDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-vo_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "用户注销登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "注销",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "接收用户账号密码进行注册, 注册成功返回用户的 id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册参数",
                        "name": "userRegisterDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-uint"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "put": {
                "description": "更新登录用户自己的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "用户更新参数",
                        "name": "userUpdateDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.R-string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.UserCreate": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "登录账号",
                    "type": "string"
                },
                "avatar": {
                    "description": "用户头像",
                    "type": "string"
                },
                "role": {
                    "description": "用户角色",
                    "type": "string"
                },
                "userName": {
                    "description": "用户昵称",
                    "type": "string"
                }
            }
        },
        "dto.UserLogin": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "登录账号",
                    "type": "string"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string"
                }
            }
        },
        "dto.UserRegister": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "登录账号",
                    "type": "string"
                },
                "checkPass": {
                    "description": "二次输入密码",
                    "type": "string"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string"
                }
            }
        },
        "dto.UserUpdate": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "用户头像",
                    "type": "string"
                },
                "id": {
                    "description": "用户 id",
                    "type": "integer"
                },
                "profile": {
                    "description": "用户简介",
                    "type": "string"
                },
                "role": {
                    "description": "用户角色",
                    "type": "string"
                },
                "userName": {
                    "description": "用户昵称",
                    "type": "string"
                }
            }
        },
        "entity.Page-vo_User": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.User"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "deleteTime": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "response.R-entity_Page-vo_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/entity.Page-vo_User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.R-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.R-uint": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.R-vo_User": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/vo.User"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "vo.User": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "deleteTime": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "profile": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}